requirements.txt
pyyaml
pydub
irc
google-cloud-texttospeech
google-cloud-core